include document focus
